requirements.txt
annotated-types==0.7.0
coverage==7.11.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.42.1
google-genai==1.48.0
openai-whisper==20250625
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-timeout==2.4.0
python-dotenv==1.2.1
tiktoken==0.12.0
torch==2.9.0
tqdm==4.67.1
typing_extensions==4.15.0
yt-dlp==2025.11.3.233024.dev0
